refractor: Moved endpoints related to another file (Post.cs)
Feat: Added Validation To create post endpoints
