+++ MySocialMedia.Api/.GITIGNORE
