+++ MySocialMedia.Api/posts.http
@@ -6,8 +6,8 @@ POST http://localhost:5036/posts
 Content-Type: application/json
 
 {
-    "username": "Jayden",
-    "title": "How To Cook",
+    # "username": "Jayden",
+    # "title": "How To Cook",
     "content": "This is a random Post that teaches how to Cookie"
 }
 
@@ -23,4 +23,4 @@ Content-Type: application/json
 
 
 ###
-DELETE http://localhost:5036/posts/3
+DELETE http://localhost:5036/posts/1
